Remove highlighting cruft.  Closes #18
- changed how the highlighting is hidden/shown; was adding/removing
the highlighting divs from the document, but that was dumb; now
just changing display from "none" to "block".

- simplified the identification in NodeStore of the highlighting
divs so they aren't sent by the client - was keeping track of
the parent div, and doing identity check, now just checking
for a property on the element.

- drive-by fix to add a thin border to the highlighting divs so
they stand out a bit more