Opt-in for not implemented messages. Closes #19
The default code path no longer logs not implemented messages
from Weinre.notImplemented().  There is now a switch to enable
them to be shown.  The switch is set in the client, as they
will only appear in the client if you debug the client web
page.  The switch is also set in the demo-pieces web page.