add back the ant code to download vendor libraries
Previously, I had a separate project, not included with weinre,
that downloaded the required vendor files from the intertubes,
zipped them up, put them on the web for weinre to download.

Changed to download those libraries in weinre's main build.

The linkage between those weinre and that other project was a bit
too loose for me.