update WebKit version, throw Ex instead of Error
updated the version of WebKit to something from today - will likely
pull more recent versions along the way.

new version of WebKit does not include InjectedScript.js, so added
some code in the build script to create it from the changed bits.

made a drive-by fix to throw Ex() instead of Error(), when I throw
errors.  Ex() throws an error but produces a best-effort stack
trace on the console before doing so.