support for client and target to handle multi-user
client and target now accept an "id" after a hash character (#)
in places where URLs are expected, like the primary client URL,
and the URL for the target script you inject.  This id will
be used at connection time to scope per-user sessions on a
server that supports multiple users.