fix https://issues.apache.org/jira/browse/CB-257
the README.md in the weinre-node distribution is bogus

For some reason, the node_modules in weinre.server got munged.  A fresh
`npm install` got everything fixed up, but apparently auto-updated
formidable.  That's not great.  Presumably it updated due to a
"use version >= x.y.z" in one of weinre's (recursive) pre-reqs.

Beyond that, did some README.md fix ups.

Also changed the `weinre` command to load coffee-script via an
explicit path; this might not be needed (the package.json file
for coffeescript was munged, but then fixed later).