fix a number of error conditions
bug: https://issues.apache.org/jira/browse/CB-284

invocation exception on WeinreClientEventsImpl.connectionCreated(): \
    TypeError: Cannot read property 'url' of undefined

The original problem logged in this issue was just one of the
problems fixed.  There were a handful of others.  Added
error checking where I could, and where I couldn't, I changed
the logger to not display a warning message (bug in Web Inspector).

Drive-by fixes for:

- changing the sequence number of channels to start at 1 instead
of a random number.  Note that the primary reason to use a
random number is to because the server can get confused
when you start/stop but have clients/targets that try to
reconnect with the same channel number.  Not a problem at all
in deployment, but a hassle at development time - or used to
be.  I'm going to try turning the 'random' off and see how
it goes.

- dump the pid at startup when in verbose mode; making it
easier to run top/dtrace/etc.