bullet-proof'd modjewel
bug: https://issues.apache.org/jira/browse/CB-293

`require()` and `define()` are no longer globally defined
in modjewel, only the object `modjewel` is globally defined.

Changed references of `define()` to `modjewel.define()`, and
change the initial launching `require()`s in the client and target
to `modjewel.require()`.  There were also a couple of 'scripts'
which needed changes to the `modjewel` specific version of `require()`.

Drive-by fix to upgrade modjewel version to 2.0, and found an
old 512x512 weinre icon (from the old mac app's .icns file) to
add in case we need bigger than 128x128